+++ src/map-object.ts
@@ -2,11 +2,9 @@ export function mapObject(
   input: Record<string, unknown>,
   mapper: (value: unknown) => unknown
 ): Record<string, unknown> {
-  return Object.entries(input).reduce(
-    (soFar, [key, value]) => ({
-      ...soFar,
-      [key]: mapper(value),
-    }),
-    {}
-  )
+  const output: Record<string, unknown> = {}
+  for (const key in input) {
+    output[key] = mapper(input[key])
+  }
+  return output
 }